新增ba crud接口

+++ tieba-adapter/src/main/java/com/tingyu/tieba/web/CommentController.java
@@ -1,5 +1,7 @@
 package com.tingyu.tieba.web;
 
+import java.math.BigInteger;
+
 import com.alibaba.cola.dto.MultiResponse;
 import com.alibaba.cola.dto.Response;
 import com.tingyu.tieba.comment.api.CommentServiceI;
@@ -21,7 +23,7 @@ public class CommentController {
     private CommentServiceI commentService;
 
     @GetMapping(value = "/comment")
-    public MultiResponse<CommentDTO> listCommentByTieId(@RequestParam(required = false) Long tieId){
+    public MultiResponse<CommentDTO> listCommentByTieId(@RequestParam(required = false) BigInteger tieId){
         CommentListByTieIdQry commentListByTieIdQry = new CommentListByTieIdQry();
         commentListByTieIdQry.setTieId(tieId);
         return commentService.listByTieId(commentListByTieIdQry);
