+++ tieba-client/src/main/java/com/tingyu/tieba/comment/dto/data/CommentDTO.java
@@ -1,5 +1,6 @@
 package com.tingyu.tieba.comment.dto.data;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import javax.validation.constraints.NotEmpty;
@@ -9,16 +10,16 @@ import lombok.Data;
 @Data
 public class CommentDTO{
 
-    private Long id;
+    private BigInteger id;
     @NotEmpty
-    private Long tieId;
-    private Long parentId;
+    private BigInteger tieId;
+    private BigInteger parentId;
     private String content;
-    private Long userId;
+    private BigInteger userId;
     private String creator;
     private LocalDateTime createTime;
     private LocalDateTime modifyTime;
-    private Long likes;
+    private BigInteger likes;
     private String status;
     private String remark;
 }
