+++ tieba-client/src/main/java/com/tingyu/tieba/comment/dto/query/CommentListByTieIdQry.java
@@ -1,9 +1,12 @@
 package com.tingyu.tieba.comment.dto.query;
 
+import java.math.BigInteger;
+
 import com.alibaba.cola.dto.Query;
+
 import lombok.Data;
 
 @Data
 public class CommentListByTieIdQry extends Query{
-   private Long tieId;
+   private BigInteger tieId;
 }
