+++ tieba-infrastructure/src/main/java/com/tingyu/tieba/mappers/BaMapper.java
@@ -1,5 +1,6 @@
 package com.tingyu.tieba.mappers;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import com.tingyu.tieba.ba.dataobject.BaDO;
@@ -10,11 +11,11 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface BaMapper{
 
-  public BigInteger create(BaDO baDO);
+  public BigDecimal create(BaDO baDO);
 
-  public BigInteger update(BaDO baDO);
+  public BigDecimal update(BaDO baDO);
 
-  public BigInteger delete(@Param("id") BigInteger id);
+  public BigDecimal delete(@Param("id") BigInteger id);
   
   public List<BaDO> list(BaListQry baListQry);
